requirements.txt
# TDS Monitoring System Requirements
streamlit==1.29.0
fastapi==0.108.0
uvicorn==0.25.0
pyserial==3.5
pandas==2.1.4
plotly==5.18.0
pyyaml==6.0.1
python-dateutil==2.8.2
requests==2.31.0